Extract privileged command wrapper into util
Without this, it makes the code a bit harder to read.

This fixes it by extracting the method.
