+++ cmd/kured/main_test.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"github.com/kubereboot/kured/pkg/util"
 	"reflect"
 	"testing"
 
@@ -22,6 +23,7 @@ func (fbc BlockingChecker) isBlocked() bool {
 
 var _ RebootBlocker = BlockingChecker{}       // Verify that Type implements Interface.
 var _ RebootBlocker = (*BlockingChecker)(nil) // Verify that *Type implements Interface.
+
 
 func Test_flagCheck(t *testing.T) {
 	var cmd *cobra.Command
@@ -107,85 +109,62 @@ func Test_stripQuotes(t *testing.T) {
 	}
 }
 
+
 func Test_rebootBlocked(t *testing.T) {
-	noCheckers := []RebootBlocker{}
-	nonblockingChecker := BlockingChecker{blocking: false}
-	blockingChecker := BlockingChecker{blocking: true}
+       noCheckers := []RebootBlocker{}
+       nonblockingChecker := BlockingChecker{blocking: false}
+       blockingChecker := BlockingChecker{blocking: true}
 
-	// Instantiate a prometheusClient with a broken_url
-	promClient, err := alerts.NewPromClient(papi.Config{Address: "broken_url"})
-	if err != nil {
-		log.Fatal("Can't create prometheusClient: ", err)
-	}
-	brokenPrometheusClient := PrometheusBlockingChecker{promClient: promClient, filter: nil, firingOnly: false}
+       // Instantiate a prometheusClient with a broken_url
+       promClient, err := alerts.NewPromClient(papi.Config{Address: "broken_url"})
+       if err != nil {
+               log.Fatal("Can't create prometheusClient: ", err)
+       }
+       brokenPrometheusClient := PrometheusBlockingChecker{promClient: promClient, filter: nil, firingOnly: false}
 
-	type args struct {
-		blockers []RebootBlocker
-	}
-	tests := []struct {
-		name string
-		args args
-		want bool
-	}{
-		{
-			name: "Do not block on no blocker defined",
-			args: args{blockers: noCheckers},
-			want: false,
-		},
-		{
-			name: "Ensure a blocker blocks",
-			args: args{blockers: []RebootBlocker{blockingChecker}},
-			want: true,
-		},
-		{
-			name: "Ensure a non-blocker doesn't block",
-			args: args{blockers: []RebootBlocker{nonblockingChecker}},
-			want: false,
-		},
-		{
-			name: "Ensure one blocker is enough to block",
-			args: args{blockers: []RebootBlocker{nonblockingChecker, blockingChecker}},
-			want: true,
-		},
-		{
-			name: "Do block on error contacting prometheus API",
-			args: args{blockers: []RebootBlocker{brokenPrometheusClient}},
-			want: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := rebootBlocked(tt.args.blockers...); got != tt.want {
-				t.Errorf("rebootBlocked() = %v, want %v", got, tt.want)
-			}
-		})
-	}
+       type args struct {
+               blockers []RebootBlocker
+       }
+       tests := []struct {
+               name string
+               args args
+               want bool
+       }{
+               {
+                       name: "Do not block on no blocker defined",
+                       args: args{blockers: noCheckers},
+                       want: false,
+               },
+               {
+                       name: "Ensure a blocker blocks",
+                       args: args{blockers: []RebootBlocker{blockingChecker}},
+                       want: true,
+               },
+               {
+                       name: "Ensure a non-blocker doesn't block",
+                       args: args{blockers: []RebootBlocker{nonblockingChecker}},
+                       want: false,
+               },
+               {
+                       name: "Ensure one blocker is enough to block",
+                       args: args{blockers: []RebootBlocker{nonblockingChecker, blockingChecker}},
+                       want: true,
+               },
+               {
+                       name: "Do block on error contacting prometheus API",
+                       args: args{blockers: []RebootBlocker{brokenPrometheusClient}},
+                       want: true,
+               },
+       }
+       for _, tt := range tests {
+               t.Run(tt.name, func(t *testing.T) {
+                       if got := rebootBlocked(tt.args.blockers...); got != tt.want {
+                               t.Errorf("rebootBlocked() = %v, want %v", got, tt.want)
+                       }
+               })
+       }
 }
 
-func Test_buildHostCommand(t *testing.T) {
-	type args struct {
-		pid     int
-		command []string
-	}
-	tests := []struct {
-		name string
-		args args
-		want []string
-	}{
-		{
-			name: "Ensure command will run with nsenter",
-			args: args{pid: 1, command: []string{"ls", "-Fal"}},
-			want: []string{"/usr/bin/nsenter", "-m/proc/1/ns/mnt", "--", "ls", "-Fal"},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			if got := buildHostCommand(tt.args.pid, tt.args.command); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("buildHostCommand() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
 
 func Test_buildSentinelCommand(t *testing.T) {
 	type args struct {
