+++ pkg/util/util.go
@@ -1,6 +1,7 @@
 package util
 
 import (
+	"fmt"
 	"os/exec"
 
 	log "github.com/sirupsen/logrus"
@@ -19,5 +20,16 @@ func NewCommand(name string, arg ...string) *exec.Cmd {
 		WithField("std", "err").
 		WriterLevel(log.WarnLevel)
 
+	return cmd
+}
+
+// PrivilegedHostCommand wraps the command with nsenter.
+// It allows to run a command from systemd's namespace for example (pid 1)
+// This relies on hostPID:true and privileged:true to enter host mount space
+// For info, rancher based need different pid, which should be user given.
+// until we have a better discovery mechanism.
+func PrivilegedHostCommand(pid int, command []string) []string {
+	cmd := []string{"/usr/bin/nsenter", fmt.Sprintf("-m/proc/%d/ns/mnt", pid), "--"}
+	cmd = append(cmd, command...)
 	return cmd
 }
